requirements.txt
mysql-connector-python==8.0.13
Scrapy==1.5.1